refactor: 컴포넌트 import 방식
