+++ apps/service/app/components/Core.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import type { PropsWithChildren } from 'react';
+import React, { type PropsWithChildren } from 'react';
+import type { Session } from 'next-auth';
+
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { SessionProvider } from 'next-auth/react';
-import React from 'react';
-import { Session } from 'next-auth';
 
 const client = new QueryClient();
