+++ apps/service/app/components/index.ts
@@ -1,0 +1,4 @@
+export { default as Core } from './Core';
+export { default as Navbar } from './Navbar';
+export { default as Search } from './Search';
+export { default as CountUp } from './CountUp';
