+++ apps/service/app/layout.tsx
@@ -4,8 +4,7 @@ import type { Metadata } from 'next';
 import { Suspense, type PropsWithChildren } from 'react';
 import { getServerSession } from 'next-auth';
 
-import Navbar from '~/components/Navbar';
-import Core from '~/components/Core';
+import { Navbar, Core } from '~/components';
 import { authOptions } from '~/api/auth/[...nextauth]/route';
 
 export const metadata: Metadata = {
